Create Odd Man Out.cpp
/*You're given an unsorted array of integers where every integer appears exactly
twice, except for one integer which appears only once. Write an algorithm (in a
language of your choice) that finds the integer that appears only once.

Good Answer: Set up a hash set that we will put the integers from the array into.
Have a second variable that will keep a sum. Start going through the array and for
each integer, check to see if it's already in the hash set. If it is not, add that integer to
the sum and store that integer in the hash set. If it is in the hash set, subtract that
integer from the sum. When the algorithm finishes going through the array, the sum
variable should be equal to the integer we were looking for, since it is the only
number we never subtracted from the sum. This takes O(n) time and O(n) space*/